Accept all actiontypes
When querying an order, LogicBoxes can return the undocumented
field "actiontype", which indicates that a previously requested action
on the order is still being processed.  The code has always recognized
a few of these actiontypes, and contructs a different class if the
actiontype indicates that a domain is still being transferred in.
However, the code throws an error on any actiontype it does not
recognize.  Unfortunately, LogicBoxes says that new actiontypes can
be added at any time.  Furthermore, there are no known actiontypes that
should prevent normal object construction.

In partituclar, the actiontypes RenewDomain and ParkDomain have been
observed when trying to reload an order immediately after requesting
renewal or deletion of a domain, and the code croaking on these has
caused errors to be reported even when no actual error has occurred.

This patch makes the code accept all actiontype values so as to avoid
croaking when it is not necessary.